requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==7.0.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9


  1 | pbkdf2_sha256$260000$7IlM3En5vK7VsD0dAIvTfH$wilzLjQEed0jZjJNhjf3BBoJIZTNT/cskqS9c2jRi8E= |            | riphlogiston@gmail.com | daria |           | t         | f        | 
(1 row)

~

  1 | !HVokh7FjcwcodMd45HjuwNj6ijIMtUEUDgjNB4HN |            | riphlogiston@gmail.com | daria |           | t         | f        | 
(1 row)